+++ src/git_fork_recon/git/repo.py
@@ -29,7 +29,7 @@ class CommitInfo:
 class GitRepo:
     def __init__(self, repo_info: RepoInfo, config: Config):
         self.repo_info = repo_info
-        self.cache_dir = config.cache_dir
+        self.cache_dir = config.cache_repo
         self.repo_dir = self.cache_dir / repo_info.owner / repo_info.name
         self._ensure_repo()